+++ docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0-preview
+### Added
+- Make some internal methods public
+### Changed
+- Move some extension methods with no reference to the extended class to the `SurfaceMapImage`
+  static class
+
 ## 0.1.0-preview
 ### Added
 - Initial preview release